renamed components and fixed broken imports
