+++ internal/ui/components/previewlist/previewlist.go
@@ -1,12 +1,12 @@
-package combined
+package previewlist
 
 import (
 	"context"
 	"log"
 
 	"github.com/Fomiller/mixify/internal/auth"
-	"github.com/Fomiller/mixify/internal/ui/components/confirm"
-	"github.com/Fomiller/mixify/internal/ui/components/track"
+	"github.com/Fomiller/mixify/internal/ui/components/textinput"
+	"github.com/Fomiller/mixify/internal/ui/components/tracklist"
 	"github.com/Fomiller/mixify/internal/ui/messages"
 	"github.com/Fomiller/mixify/internal/ui/styles"
 	"github.com/charmbracelet/bubbles/key"
@@ -104,7 +104,7 @@ func (m Model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 
 func (m Model) View() string {
 	if m.Confirm == true {
-		cm := confirm.New()
+		cm := textinput.New()
 		return styles.DocStyle.Render(cm.View())
 	}
 	h, _ := styles.DocStyle.GetFrameSize()
@@ -136,7 +136,7 @@ func (m Model) CreatePlaylist(name string, desc string) error {
 	var trackIDs []spotify.ID
 	tracks := m.List.Items()
 	for _, t := range tracks {
-		x := t.(track.Item)
+		x := t.(tracklist.Item)
 		trackIDs = append(trackIDs, x.TrackID)
 	}
 	// make multiple calls to add tracks if needed, spotify only supports 100 at a time
